+++ src/cmd/list/mod.rs
@@ -2,6 +2,7 @@ use clap::Parser;
 
 pub(super) mod accounts;
 pub(super) mod api_keys;
+pub(super) mod generation_features;
 pub(super) mod issue_comments;
 pub(super) mod issues;
 pub(super) mod linked_accounts;
@@ -30,6 +31,9 @@ pub enum Subcommand {
 
     #[command(name = "api-keys", alias = "apiKeys")]
     ApiKeys(api_keys::Opts),
+
+    #[command(name = "generation-features", alias = "generationFeatures")]
+    GenerationFeatures(generation_features::Opts),
 
     #[command(name = "issue-comments", alias = "issueComments")]
     IssueComments(issue_comments::Opts),
@@ -63,6 +67,9 @@ pub async fn run(subcommand: Subcommand) -> Result<(), Box<dyn std::error::Error
         }
         Subcommand::ApiKeys(api_keys::Opts { offset, limit, output }) => {
             api_keys::run(offset, limit, output).await?;
+        }
+        Subcommand::GenerationFeatures(generation_features::Opts { offset, limit, output }) => {
+            generation_features::run(offset, limit, output).await?;
         }
         Subcommand::IssueComments(issue_comments::Opts { offset, limit, output }) => {
             issue_comments::run(offset, limit, output).await?;