+++ build.rs
@@ -1,9 +1,0 @@
-use anyhow::Result;
-
-fn main() -> Result<()> {
-    // Configure vergen
-    let build = vergen_gitcl::GitclBuilder::all_git()?;
-    vergen_gitcl::Emitter::default().add_instructions(&build)?.emit()?;
-
-    Ok(())
-}